+++ backend/canisters/storage_bucket/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Avoid using `candid::Func` type directly ([#3983](https://github.com/open-chat-labs/open-chat/pull/3983))
+
 ## [[2.0.717](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.717-storage_bucket)] - 2023-06-12
 
 ### Changed